Add environment switching; add default repo switch
This commit adds the NODE_ENV environment variable.

Current acceptable values are ``test`` and ``dev``. If ``test`` is
selected, the configuration will be loaded from ``test/config.json``.

This commit also adds a configuration option to change the repositories
listed on the index page by default.

The .gitignore has been edited to allow test/config.json to be
committed, which contains a standardised set of repositories that will
remain static and therefore useful for testing.

The travis build configuration was also updated to set the ``test``
environment.